refactor(direct-control): extract action approval primitive
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves Civ7ActionApproval and the shared assertApproved helper into src/action-approval.ts while preserving public facade type re-exports through src/play/operations/types.ts and existing facade dependency injection.

This preserves approval-first rejection message/classification and adds focused package proof that a blank approval reason rejects before socket use. It does not define operation telemetry, AI ingestion, semantic CLI projection, hotseat runtime proof, Effect/Bun resource/layer/schedule composition, or Effect/oRPC procedure-core work. Task 2.9.4 matrix-row acceptance remains pending.

Validation:
- bun run --cwd packages/civ7-direct-control test -- unit-operation.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- git diff --cached --check

Local package/source relocation proof only; no runtime, live-game, AI-on-hotseat, telemetry, semantic CLI, Effect/oRPC, or procedure-core proof claimed.
